requirements.txt
discord.py[voice]
ffmpeg
youtube-dl
